+++ src/lib/error/ZWaveError.ts
@@ -26,6 +26,11 @@ export enum ZWaveErrorCodes {
 	Controller_InclusionFailed,
 	Controller_ExclusionFailed,
 
+	/** The node with the given node ID was not found */
+	Controller_NodeNotFound,
+	/** The endpoint with the given index was not found on the node */
+	Controller_EndpointNotFound,
+
 	CC_Invalid,
 	CC_NoNodeID,
 	CC_NotSupported,
@@ -64,6 +69,19 @@ export enum ZWaveErrorCodes {
 	 * Used to report that the command was not executed by the target node
 	 */
 	SupervisionCC_CommandFailed = 1100,
+
+	/**
+	 * Used to report that a ManufacturerProprietaryCC could not be instanciated
+	 * because of a missing manufacturer ID.
+	 */
+	ManufacturerProprietaryCC_NoManufacturerId = 1200,
+
+	/**
+	 * Used to report that an invalid group ID was used to address a (Multi Channel) Association
+	 */
+	AssociationCC_InvalidGroup = 1300,
+	/** Cannot add an association because it is not allowed */
+	AssociationCC_NotAllowed,
 }
 
 /**