+++ CHANGELOG.md
@@ -4,18 +4,41 @@
 	## __WORK IN PROGRESS__
 -->
 
-## __WORK IN PROGRESS__
+## 2.15.5 (2020-03-15)
+### Bugfixes
+* Delayed the check for manufacturer ID until the `Manufacturer Proprietary CC` needs it to avoid crashing during cache serialization
+
+## 2.15.4 (2020-03-11)
+### Bugfixes
+* The `FibaroVenetianBlindCCGet` now correctly expects a response
+
+## 2.15.3 (2020-03-09)
+### Bugfixes
+* The `firmwareVersion` property of a node now returns a value again
+* Fixed the interview procedure of `Manufacturer Proprietary CC` by reading the manufacturer ID from the Value DB inside the constructor
+
+## 2.15.2 (2020-03-07)
+### Bugfixes
+* Fixed the logic for filtering out root endpoint values
+
+## 2.15.1 (2020-03-07)
+### Bugfixes
+* Send data transmit reports for singlecast messages are detected correctly again
+
+## 2.15.0 (2020-03-07)
 ### Featues
 * Config files may now specify manufacturer proprietary parameters. This can be used to enable certain manufacturer proprietary commands
 * Completed support for the `Fibaro Venetian Blind CC`
 * Added support for some legacy devices by implementing `Multi Instance CC` (version 1 of the `Multi Channel CC`)
 * When the wake up interval of a device seems to be longer than configured, the current interval is now re-queried
 * Upon receipt of a `ClockCCReport` which deviates from the controller's clock, the sending node's clock setting now gets updated
+* Value IDs of the root endpoint which have a corresponding value on another endpoint are now filtered out
 
 ### Bugfixes
 * Fixed a compilation issue regarding `Send Data` message arguments
 * When testing potential responses of Multi Channel requests, the response's source endpoint is now checked
 * When testing potential responses of `ConfigurationGet` requests, the response's parameter number is now checked
+* The device config for the controller node now gets loaded if possible
 
 ### Configuration updates
 * Updated `ZHC5002` configuration for firmware versions >= 2.02